# docs/api-spec.json
{
    "swagger": "2.0",
    "info": {
      "description": "Deel is a Global Payroll and Compliance for Employees and Contractors Around the World all in one click.",
      "version": "1.0.0",
      "title": "Deel",
      "termsOfService": "http://swagger.io/terms/",
      "contact": {
        "email": "apiteam@swagger.io"
      },
      "license": {
        "name": "Apache 2.0",
        "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
      }
    },
    "host": "localhost:3001",
    "tags": [
      {
        "name": "profiles",
        "description": "A profile can be either a `client` or a `contractor`. Clients create contracts with contractors. contractor does jobs for clients and get paid. Each profile has a balance property."
      },
      {
        "name": "contracts",
        "description": "Contracts have 3 statuses, `new`, `in_progress`, `terminated`. contracts are considered active only when in status `in_progress`. Contracts group jobs within them."
      },
      {
        "name": "jobs",
        "description": "Contractor get paid for jobs by clients under a certain contract."
      }
    ],
    "schemes": [
      "http"
    ],
    "paths": {
      "/admin/best-profession": {
        "get": {
          "tags": [
            "profiles"
          ],
          "description": "Returns the profession that earned the most money (sum of jobs paid) for any contactor that worked in the query time range.",
          "operationId": "getBestProfession",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "profile_id",
              "in": "header",
              "type": "string",
              "required": true
            },
            {
              "name": "start",
              "in": "query",
              "required": true,
              "type": "string"
            },
            {
              "name": "end",
              "in": "query",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "204": {
              "description": "No Content"
            },
            "400": {
              "description": "Bad Request"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/admin/best-clients": {
        "get": {
          "tags": [
            "profiles"
          ],
          "description": "Returns the clients the paid the most for jobs in the query time period. limit query parameter should be applied, default limit is 2.",
          "operationId": "getBestClients",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "profile_id",
              "in": "header",
              "type": "string",
              "required": true
            },
            {
              "name": "start",
              "in": "query",
              "required": true,
              "type": "string"
            },
            {
              "name": "end",
              "in": "query",
              "required": true,
              "type": "string"
            },
            {
              "name": "limit",
              "in": "query",
              "required": true,
              "type": "number"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "204": {
              "description": "No Content"
            },
            "400": {
              "description": "Bad Request"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/balances/deposit/{userId}": {
        "post": {
          "tags": [
            "profiles"
          ],
          "description": "Deposits money into the the the balance of a client, a client can't deposit more than 25% his total of jobs to pay. (at the deposit moment)",
          "operationId": "doDeposit",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "profile_id",
              "in": "header",
              "type": "string",
              "required": true
            },
            {
              "name": "userId",
              "in": "path",
              "description": "ID that need to be updated",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "body",
              "description": "Updated class object",
              "required": true,
              "schema": {
                "$ref": "#/definitions/doDeposit"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "204": {
              "description": "No Content"
            },
            "400": {
              "description": "Bad Request"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/contracts": {
        "get": {
          "tags": [
            "contracts"
          ],
          "description": "Returns a list of contracts belonging to a user (client or contractor), the list should only contain non terminated contracts.",
          "operationId": "getAll",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "profile_id",
              "in": "header",
              "type": "string",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "204": {
              "description": "No Content"
            },
            "400": {
              "description": "Bad Request"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/contracts/{id}": {
        "get": {
          "tags": [
            "contracts"
          ],
          "description": "Return the contract only if it belongs to the profile calling.",
          "operationId": "getById",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "profile_id",
              "in": "header",
              "type": "string",
              "required": true
            },
            {
              "name": "id",
              "in": "path",
              "type": "string",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "204": {
              "description": "No Content"
            },
            "400": {
              "description": "Bad Request"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/jobs/unpaid": {
        "get": {
          "tags": [
            "jobs"
          ],
          "description": "Get all unpaid jobs for a user (***either*** a client or contractor), for ***active contracts only***.",
          "operationId": "getUnpaid",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "profile_id",
              "in": "header",
              "type": "string",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "204": {
              "description": "No Content"
            },
            "400": {
              "description": "Bad Request"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/jobs/{job_id}/pay": {
        "post": {
          "tags": [
            "jobs"
          ],
          "description": "Pay for a job, a client can only pay if his balance >= the amount to pay. The amount should be moved from the client's balance to the contractor balance.",
          "operationId": "doPayment",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "profile_id",
              "in": "header",
              "type": "string",
              "required": true
            },
            {
              "name": "job_id",
              "in": "path",
              "type": "string",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "204": {
              "description": "No Content"
            },
            "400": {
              "description": "Bad Request"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      }
    },
    "definitions": {
      "doDeposit": {
        "type": "object",
        "required": [
          "deposit"
        ],
        "properties": {
          "deposit": {
            "type": "number",
            "example": 10
          }
        }
      }
    },
    "externalDocs": {
      "description": "Find out more about Swagger",
      "url": "http://swagger.io"
    }
  }